refactor: extract hook from component
